convert rename and pag number dialogs. update keyboard shortcut handling

+++ frontend/src/views/Unit.vue
@@ -1,16 +1,6 @@
 <template>
    <div class="unit">
       <WaitSpinner  v-if="unitStore.working" :overlay="true" message="Working..." />
-      <ConfirmDialog group="delete">  <!-- FIXME -->
-         <template #message>
-            <div style="display:flex; flex-direction: column; gap: 10px; align-items: flex-start;">
-               <div>Delete the selected images? All data will be lost.</div>
-               <div>This is not reversable.</div>
-               <div>Are you sure?</div>
-            </div>
-         </template>
-      </ConfirmDialog>
-
       <div class="metadata" v-if="projectStore.hasDetail">
          <h2>
             <ProblemsDisplay class="topleft" />
@@ -19,7 +9,6 @@
          <h3>
             <div>{{callNumber}}</div>
             <div>Unit {{unitStore.unitID}}</div>
-            <div class="divider"></div>
             <div class="small" >{{workingDir}}</div>
             <div class="small" >{{unitStore.masterFiles.length}} Images</div>
          </h3>
@@ -40,19 +29,25 @@ import ProblemsDisplay from '@/components/ProblemsDisplay.vue'
 import {useProjectStore} from "@/stores/project"
 import {useSystemStore} from "@/stores/system"
 import {useUnitStore} from "@/stores/unit"
-import { computed, onMounted, onBeforeUnmount } from 'vue'
+import { computed, onMounted } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import MasterFilesList from '@/components/unit/MasterFilesList.vue'
 import MasterFilesGrid from '@/components/unit/MasterFilesGrid.vue'
-import { useConfirm } from "primevue/useconfirm"
 import KeyboardShortcutHelp from '@/components/KeyboardShortcutHelp.vue'
+import { onKeyStroke } from '@vueuse/core'
+import { useConfirm } from "../composables/useConfirm"
 
 const projectStore = useProjectStore()
 const systemStore = useSystemStore()
 const unitStore = useUnitStore()
 const route = useRoute()
 const router = useRouter()
-const confirm = useConfirm()
+
+onKeyStroke('d', (e) => {
+   if ( e.ctrlKey ) {
+      handleDelete()
+   }
+})
 
 const title = computed(() => {
    let t = projectStore.detail.title
@@ -92,56 +87,20 @@ function paddedUnit() {
    return unitStr.padStart(9,'0')
 }
 
-const handleDelete = (() => {
-   confirm.require({
-      group: 'delete',
-      header: 'Confirm Image Delete',
-      accept: () => {
-         unitStore.deleteSelectedMasterFiles()
-      }
-   })
-})
-
-const keyboardHandler = ((event) => {
-   if ( event.target.id == "edit-desc" || event.target.id == "title-input-box" ||
-        unitStore.edit.pageNumber || unitStore.edit.metadata || unitStore.edit.component ) {
-      return
-   }
-
-   if ( event.key == ',' || event.key == '<') {
-      if (unitStore.currPage > 1) {
-         priorClicked()
-         return
-      }
-   }
-   if ( event.key == '.' || event.key == '>') {
-      if (unitStore.currPage < unitStore.totalPages) {
-         nextClicked()
-         return
-      }
-   }
-
-   if ( !event.ctrlKey ) return
-
-   if (event.key == 'p') {
-      unitStore.edit.pageNumber = true
-   } else if (event.key == 'b' || event.key == 'f') {
-      unitStore.edit.metadata = true
-   } else if (event.key == 'k') {
-      unitStore.edit.component = true
-   }  else if (event.key == 'a') {
-      unitStore.selectPage()
-   }  else if (event.key == 'd') {
-      handleDelete()
-   }
+const handleDelete = ( async () => {
+   let msg = `<div style="display:flex; flex-direction: column; gap: 5px; align-items: flex-start;">
+               <div>Delete the selected images? All data will be lost.</div>
+               <div>This is not reversable.</div>
+               <div>Are you sure?</div>
+            </div>`
+   const resp = await useConfirm("Confirm Image Delete", msg, "Delete")
+   if (resp) {
+      unitStore.deleteSelectedMasterFiles()
+   } 
 })
 
 onMounted( async () => {
    unitStore.lastURL = ""
-
-   // setup keyboard litener for shortcuts
-   window.addEventListener('keyup', keyboardHandler)
-   // window.addEventListener("scroll", scrollHandler)
 
    if (projectStore.hasDetail == false) {
       await projectStore.getProject(route.params.id)
@@ -164,10 +123,6 @@ onMounted( async () => {
    }
 })
 
-onBeforeUnmount( async () => {
-   window.removeEventListener('keyup', keyboardHandler)
-})
-
 </script>
 
 <style lang="scss" scoped>
@@ -175,15 +130,11 @@ onBeforeUnmount( async () => {
    padding: 0;
    text-align: center;
 
-   input[type=checkbox] {
-      width: 20px;
-      height: 20px;
-   }
    .metadata {
       margin-bottom: 15px;
       position: relative;
       .small {
-         font-size: 0.85em;
+         font-size: 0.95em;
          margin-bottom: 5px;
       }
       .topleft {
@@ -203,28 +154,13 @@ onBeforeUnmount( async () => {
          }
       }
       h3 {
-          margin: 5px 0 25px 0;
+          margin: 5px 0;
           font-weight: normal;
           .divider {
             border-bottom: 1px solid var(--uvalib-grey-light);
             margin: 10px auto 20px auto;
             width: 50%;
           }
-      }
-      a {
-         display: inline-block;
-         margin-top: 8px;
-         font-weight: bold;
-         text-decoration: none;
-         cursor: pointer;
-         color: var(--uvalib-blue-alt);
-         .link {
-            display: inline-block;
-            margin-left: 8px;
-         }
-         &:hover {
-            text-decoration: underline;
-         }
       }
       .back {
          padding: 0 15px;
